vp9_encodeframe.c: Silence totalrate assignment warning with a cast.
Cast result of round() to int.

+++ vp9/encoder/vp9_encodeframe.c
@@ -714,7 +714,7 @@ static void rd_pick_sb_modes(VP9_COMP *cpi, const TileInfo *const tile,
     x->rdmult = orig_rdmult;
     if (*totalrate != INT_MAX) {
       vp9_clear_system_state();
-      *totalrate = round(*totalrate * rdmult_ratio);
+      *totalrate = (int)round(*totalrate * rdmult_ratio);
     }
   }
   else if (cpi->oxcf.aq_mode == COMPLEXITY_AQ) {
